revert

+++ ivy_tests/test_ivy/test_frontends/test_numpy/test_mathematical_functions/test_extrema_finding.py
@@ -308,7 +308,7 @@ def test_numpy_nanmax(
 @handle_frontend_test(
     fn_tree="numpy.fmax",
     dtype_and_inputs=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
+        available_dtypes=helpers.get_dtypes("float"),
         num_arrays=2,
         min_value=-np.inf,
         max_value=np.inf,
